% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabage_data.R
\docType{data}
\name{crabage_data}
\alias{crabage_data}
\title{crabage_data}
\format{
A data frame with **rows** rows and **columns** variables:
\describe{
  \item{Weight}{Description of the weight of crabs}
  \item{Age}{Description of the age of crabs}
  ...
}
}
\usage{
crabage_data
}
\description{
This dataset contains measurements of crabs, used for predicting their age.
}
\keyword{datasets}
